+++ insteon_mqtt/device/Dimmer.py
@@ -432,28 +432,26 @@ class Dimmer(Base):
             on_done(False, "Dimmer %s state update failed", None)
 
     #-----------------------------------------------------------------------
-    def handle_group_cmd(self, addr, msg):
+    def handle_group_cmd(self, addr, group, cmd):
         """Respond to a group command for this device.
 
-        This is called when this device is a responder to a scene.
-        The device should look up the responder entry for the group in
-        it's all link database and update it's state accordingly.
+        This is called when this device is a responder to a scene.  The
+        device should look up the responder entry for the group in it's all
+        link database and update it's state accordingly.
 
         Args:
           addr:  (Address) The device that sent the message.  This is the
                  controller in the scene.
-          msg:   (message.InpStandard) The broadcast message that was sent.
-                 Use msg.group to find the scene group that was broadcast.
+          group: (int) The group being triggered.
+          cmd:   (int) The command byte being sent.
         """
         # Make sure we're really a responder to this message.  This
         # shouldn't ever occur.
-        entry = self.db.find(addr, msg.group, is_controller=False)
+        entry = self.db.find(addr, group, is_controller=False)
         if not entry:
             LOG.error("Dimmer %s has no group %s entry from %s", self.addr,
-                      msg.group, addr)
+                      group, addr)
             return
-
-        cmd = msg.cmd1
 
         # 0x11: on, 0x12: on fast
         if cmd in Dimmer.on_codes: