+++ insteon_mqtt/device/Switch.py
@@ -349,7 +349,7 @@ class Switch(Base):
             on_done(False, "Scene trigger failed failed", None)
 
     #-----------------------------------------------------------------------
-    def handle_group_cmd(self, addr, msg):
+    def handle_group_cmd(self, addr, group, cmd):
         """Respond to a group command for this device.
 
         This is called when this device is a responder to a scene.
@@ -359,28 +359,27 @@ class Switch(Base):
         Args:
           addr:  (Address) The device that sent the message.  This is the
                  controller in the scene.
-          msg:   (message.InpStandard) The broadcast message that was sent.
-                 Use msg.group to find the scene group that was broadcast.
+          group: (int) The group being triggered.
+          cmd:   (int) The command byte being sent.
         """
         # Make sure we're really a responder to this message.  This
         # shouldn't ever occur.
-        entry = self.db.find(addr, msg.group, is_controller=False)
+        entry = self.db.find(addr, group, is_controller=False)
         if not entry:
             LOG.error("Switch %s has no group %s entry from %s", self.addr,
-                      msg.group, addr)
+                      group, addr)
             return
 
         # 0x11: on, 0x12: on fast
-        if msg.cmd1 in Switch.on_codes:
+        if cmd in Switch.on_codes:
             self._set_is_on(True)
 
         # 0x13: off, 0x14: off fast
-        elif msg.cmd1 in Switch.off_codes:
+        elif cmd in Switch.off_codes:
             self._set_is_on(False)
 
         else:
-            LOG.warning("Switch %s unknown group cmd %#04x", self.addr,
-                        msg.cmd1)
+            LOG.warning("Switch %s unknown group cmd %#04x", self.addr, cmd)
 
     #-----------------------------------------------------------------------
     def _set_is_on(self, is_on):