+++ insteon_mqtt/handler/__init__.py
@@ -42,4 +42,5 @@ from .ModemDbModify import ModemDbModify
 from .ModemLinkComplete import ModemLinkComplete
 from .ModemLinkStart import ModemLinkStart
 from .ModemReset import ModemReset
+from .ModemScene import ModemScene
 from .StandardCmd import StandardCmd