+++ insteon_mqtt/mqtt/config.py
@@ -12,11 +12,14 @@ with it.
 
 #===========================================================================
 from .. import device
+from ..Modem import Modem
 from .BatterySensor import BatterySensor
 from .Dimmer import Dimmer
 from .FanLinc import FanLinc
+from .IOLinc import IOLinc
 from .KeypadLinc import KeypadLinc
 from .Leak import Leak
+from .Modem import Modem as MqttModem
 from .Motion import Motion
 from .Remote import Remote
 from .SmokeBridge import SmokeBridge
@@ -24,9 +27,11 @@ from .Switch import Switch
 
 # Map Insteon device classes to MQTT classes.
 devices = {
+    Modem : MqttModem,
     device.BatterySensor : BatterySensor,
     device.Dimmer : Dimmer,
     device.FanLinc : FanLinc,
+    device.IOLinc : IOLinc,
     device.KeypadLinc : KeypadLinc,
     device.Leak : Leak,
     device.Motion : Motion,